Refactor Rectangle and Position classes to use appropriate data types; change Rectangle constructor and properties to use int16_t and uint16_t for better precision

+++ include/neo/component/rectangle.hpp
@@ -18,13 +18,13 @@ namespace jsi::neo
 class Rectangle : public Component
 {
 public:
-    Rectangle(uint8_t x, uint8_t y, uint8_t width, uint8_t height, Color color, uint8_t z_index = 0) :
+    Rectangle(int16_t x, int16_t y, uint16_t width, uint16_t height, Color color, uint8_t z_index = 0) :
         Component(z_index)
     {
-        properties_.set<uint8_t>("x", x);
-        properties_.set<uint8_t>("y", y);
-        properties_.set<uint8_t>("width", width);
-        properties_.set<uint8_t>("height", height);
+        properties_.set<int16_t>("x", x);
+        properties_.set<int16_t>("y", y);
+        properties_.set<uint16_t>("width", width);
+        properties_.set<uint16_t>("height", height);
 
         properties_.set<uint8_t>("r", color.r_);
         properties_.set<uint8_t>("g", color.g_);
@@ -34,10 +34,10 @@ public:
 
     std::pair<Position, PixelBuffer> render() override
     {
-        auto x      = properties_.get<uint8_t>("x").value();
-        auto y      = properties_.get<uint8_t>("y").value();
-        auto width  = properties_.get<uint8_t>("width").value();
-        auto height = properties_.get<uint8_t>("height").value();
+        auto x      = properties_.get<int16_t>("x").value();
+        auto y      = properties_.get<int16_t>("y").value();
+        auto width  = properties_.get<uint16_t>("width").value();
+        auto height = properties_.get<uint16_t>("height").value();
         auto color  = Color(properties_.get<uint8_t>("r").value(),
                            properties_.get<uint8_t>("g").value(),
                            properties_.get<uint8_t>("b").value(),
